fixes #638
ok hopefully this doesn't break puzzlescript anywhere! I've checked online to see how reliable "visibilitychange" is and it seems to be generally trusted. I've checked on chrome, firefox, and edge on my pc.

I'm going to update the site right away. If anyone notices any problems, lemme know and I'll fix it asap (i worry about the awake event not getting reliably called and the game just never updating again).  (there are other less-likely-to-outright-break-if-browsers-do-things-that-they-shouldn't ways of doing this that don't give as extreme power savings. )